add encodedData field to RecordsWrite in message storage under a threshold
In this commit I've added utilized the existing RecordsWriteMessageWithOptionalEncodedData class
to leverage writing data from RecordsWrite messages directly into the MessageStore when the data
is below the 50k byte threshold.

For now I'm still storing in DataStore, but that will be removed in subsequent commits.

I had to take care and remove `encodedData` from messages when computing the CID and made sure to use the
Message.getCid() helper for computing CIDs for message storage.

Additionally, I've made sure to prune the original RecordsWrite of any encodedData when cleanup happens.