Export Sort and Pagination types.
Lower max maxDataSizeAllowedToBeEncoded. We are currently checking this against the dataSize, which is the binary data size.
However when we store it we are base64 encoding the data, which has an approximate increase of 33%, this was causing some issues going right above the limit in SQL.

Upon some rough calculations of the 65k limit for most SQL - 33% reduction, I rounded down to the nice number of 42_000.