Introduce Dwn.handleRecordsWrite function
This commit introduces a handleRecordsWrite function meant to be used in place of the existing processMessage which is eventually meant to be replaced as described in issue #289.

This commit also updates the relevant tests in records-(write|delete|read|query).ts files, which were using processMessage and replaces its usage with the new handleRecordsWrite.

It also introduces a few new tests in dwn.spec.ts, enough to bring uncovered lines of this function to 0 (more specifically, the new test is meant to cover the exception path).