EventLog / MessageStore cursor should not be a `messageCID` (#638)
Prior to this PR we used messageCid as a cursor for paginating the MessageStore, as well as EventLog.

This generally works well since all messages throughout the system can be identified by a messageCid, however it becomes tricky if a messageCid which was returned as a pagination cursor, was no longer in the MessageStore by the time the subsequent call came in using that messageCid as a cursor.

To remedy this the cursor is now a string which represents the value and itemId of the last item returned from a query. The value is the value of the sortProperty key in the item's indexes.