EventsSubscribe (#658)
This PR introduces an `EventStream` interface and an implementation based on `EventEmitter`  which emits events for any interfaces which are saved as a part of the DWN's message store. (ie. not Query/Read/Subscribe messages).

We also introduce an `EventsSubscribe` interface that follows the same authorization model as `EventsGet` and `EventsQuery`, which only allows access to the tenant owner.

In a subsequent PR we will introduce `RecordsSubscribe`, as well as some additional enhancements.

Co-authored-by: Liran Cohen <c.liran.c@gmail.com>
Co-authored-by: Andor Kesselman <andor@henosisknot.com>