Emit initialWrite with a `RecordsDelete` or a `RecordsWrite` which is updated. (#680)
This PR makes a few changes to how we emit events related to records.

#### When emitting record messages, we add an `initialWrite` property to
the event for `RecordDelete` messages and `RecordsWrite` messages which
are not the initial write.
This mimics the behavior of query when subscribing to a record, allowing
the consumer to make use of the record state which they are consuming.
If they'd like to store the record of an update or delete, they need the
initial record write along with it.

#### We only emit records which are the latest base state.
Now that we've allowed writing records without any data associated,
there are cases where a record is written but it is not queryable due to
it not existing the latest state. We don't want to emit those records as
they're useless to the consumer. The consumer will receive an event when
the subsequent record that is the latest state is written.