Signing and verifying with Secp256r1 keys (#715)
Added support for signing and verifying keys with Secp256r1 curve.
Oracle KMS utilizes these keys and we needed the verification piece for
our implementation, so thought we would share with the TBD team how we
went about it and extend to cover the signing piece as well.

Uses @noble/curves/p256 package in approach similar to
utils/secp256k1.ts. I have included tests with 100% code coverage of the
secp256r1.ts file and also tested locally with the generated key pairs
for end-to-end signing and verifying in the context of our
implementation. Open to any suggestions for improving the
implementation/excited to hear your thoughts!