RecordsPermissionScope Grants must be scoped to a protocol (#750)
Modifies `RecordsPermissionScope` to require scoping to a specific
protocol. This also removes the ability to scope to a schema as you
would just scope to a protocolPath instead.

The motivation behind this was to make querying for all Messages tied to
a specific protocol for subset sync. Without this change getting grants,
and more importantly the grant revocations which have unrestricted
scopes would be challenging/not performant.

Since we are planning on getting rid of flat-spaced records anyway, this
is a step in the right direction.

- Remove `schema` from `RecordsPermissionScope`
- `RecordsPermissionScope` requires `protocol` to be set
- The `protocol` provided within the `RecordsPermissionScope` is also
added as a `protocol` tag to the `RecordsWriteMessage` descriptor.
- When validating the Scope of a Grant `RecordsWriteMessage`, ensure the
`protocol` tag on the record is the same as the scope `protocol` defined
in the `RecordsPermissionScope`.


`grant` case is covered by this PR.