Add testing util to poll for async event testing (#758)
Due to the async nature of testing `EventStream` and Subscriptions, and
some external pub/sub systems taking a longer time to publish and/or
receive a message, we have added a helper method to help wait for a test
case to assert.

Additionally we modified the tests to be able to more accurately check
negative test cases where testing the absence of something which arrives
async could be challenging.

- `pollUntilSuccessOrTimeout` will execute a delegate function passed to
it, if it fails it will sleep `retrySleep` ms, and will retry. If the
total retry time exceeds `timeout`, it will finally fail.