Hacktoberfest 2024 Add take action to auto assign contributors (#809)
This pull request introduces a new GitHub Actions workflow to
automatically assign issues to contributors when they comment during the
month of October. The workflow also logs when the action is skipped
outside of October.

### New GitHub Actions workflow:

* `.github/workflows/take.yml`: Added a workflow named "Auto-assign
issue to contributor" that triggers on issue comments. It includes two
jobs:
*
[`assign`](diffhunk://#diff-d3e004796efd8aa465cf8563e767fe5d2f555267902130b8fe4565df1bc1a77fR1-R26):
Assigns the issue to the contributor if the comment is made in October.
*
[`log_out_of_october`](diffhunk://#diff-d3e004796efd8aa465cf8563e767fe5d2f555267902130b8fe4565df1bc1a77fR1-R26):
Logs a message indicating the action was skipped if the comment is made
outside of October.

---------

Co-authored-by: Tania Chakraborty <tchakraborty@BLKG9TC7C9CQ1.local>